requirements.txt
PyPDF2==2.12.1
Pillow==9.4.0